code/deep_da/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
certifi=2019.9.11=py36_0
cffi=1.13.2=py36hb32ad35_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=pypi_0
freetype=2.10.0=h563cfd7_1
intel-openmp=2019.4=245
jpeg=9c=hfa6e2cd_1001
kiwisolver=1.1.0=pypi_0
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
liblapack=3.8.0=14_mkl
libpng=1.6.37=h7602738_0
libtiff=4.1.0=h2e92f26_0
lz4-c=1.8.3=he025d50_1001
matplotlib=3.1.0=pypi_0
mkl=2019.4=245
ninja=1.9.0=h1ad3211_1
numpy=1.17.3=py36hc71023c_0
olefile=0.46=py_0
pandas=0.25.1=pypi_0
pillow=6.2.1=py36h643dfcc_0
pip=19.3.1=py36_0
pycparser=2.19=py36_1
pyparsing=2.4.5=pypi_0
python=3.6.7=he025d50_1005
python-dateutil=2.8.1=pypi_0
pytorch=1.3.1=py3.6_cuda101_cudnn7_0
pytz=2019.3=pypi_0
setuptools=41.6.0=py36_1
six=1.13.0=py36_0
tk=8.6.9=hfa6e2cd_1003
torchvision=0.4.2=py36_cu101
tqdm=4.31.1=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_0
wheel=0.33.6=py36_0
wincertstore=0.2=py36_1003
xz=5.2.4=h2fa13f4_1001
zlib=1.2.11=h2fa13f4_1006
zstd=1.4.3=hd8a0e53_0
